pyproject.toml
[project]
name = "who-sang-that"
version = "0.1.0"
description = "Identify who said that when?"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.12.0",
    "fastapi>=0.116.2",
    "jigsawstack>=0.3.6",
    "numpy>=2.3.3",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]
